pyproject.toml
[project]
name = "feishu-bot-sdk"
version = "0.1.0"
description = "Reusable SDK for Feishu bot messaging, Bitable, Docx, and Drive permissions."
readme = "README.md"
authors = [
    { name = "Anderson", email = "andersonby@163.com" }
]
requires-python = ">=3.10"
dependencies = [
    "betterproto2>=0.9.0",
    "httpx>=0.28.1",
    "pycryptodome>=3.20.0",
    "websockets>=15.0.1",
]

[build-system]
requires = ["uv_build>=0.9.27"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.15.4",
    "ty>=0.0.19",
]

[tool.ruff]
target-version = "py310"

[tool.ty.environment]
python = "./.venv"
